mlrest/requirements.txt
attrs==19.3.0
autopep8==1.5.3
click==7.1.2
fastapi==0.58.0
gunicorn==20.0.4
h11==0.9.0
httptools==0.1.1
importlib-metadata==1.7.0
joblib==0.15.1
mirakuru==2.3.0
more-itertools==8.4.0
numpy==1.18.5
onnx==1.7.0
onnxconverter-common==1.7.0
onnxruntime==1.3.0
packaging==20.4
pluggy==0.13.1
port-for==0.4
protobuf==3.12.2
psutil==5.7.0
py==1.9.0
pyaml==20.4.0
pycodestyle==2.6.0
pydantic==1.5.1
pyparsing==2.4.7
pytest==5.4.3
pytest-asyncio==0.14.0
pytest-mock==3.1.1
PyYAML==5.3.1
redis==3.5.3
scikit-learn==0.23.1
scipy==1.4.1
six==1.15.0
skl2onnx==1.7.0
starlette==0.13.4
threadpoolctl==2.1.0
toml==0.10.1
typing==3.7.4.1
typing-extensions==3.7.4.2
uvicorn==0.11.5
uvloop==0.14.0
wcwidth==0.2.5
websockets==8.1
zipp==3.1.0
